Introduce Yara::Scanner API This new API provides a bit more control by letting you start and stop Yara manually and by giving an API to the rule compiler and scanner that can be called separately from Yara.start and Yara.stop.
Initializing a new scanner, adding a rule or rules, and then compiling those rules
so that the scanner can use them will initialize a compiler under the hood and
then discard it once the rules are compiled.

When you are done using the scanner you must call close on it to free up the memory
used by the compiled rules.